% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_helpers.R
\name{assert_valid_par_model_spec}
\alias{assert_valid_par_model_spec}
\title{validate a par_model_spec object}
\usage{
assert_valid_par_model_spec(spec, .call = parent.frame())
}
\arguments{
\item{spec}{a par_model_spec object}

\item{.call}{parent frame}
}
\description{
validate a par_model_spec object
}
