% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nar_model_likelihood.R
\name{beta_mat}
\alias{beta_mat}
\title{Beta Matrix}
\usage{
beta_mat(betas, t)
}
\arguments{
\item{betas}{Numeric vector of beta values.}

\item{t}{Integer number of time points.}
}
\value{
Matrix of beta values.
}
\description{
Creates a matrix of beta values for a given number of time points (t) and lag values (q).
}
