% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_set_raw}
\alias{data_set_raw}
\title{Data set raw}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1798 rows and 237 columns.
}
\usage{
data_set_raw
}
\description{
Data as provided by \url{https://archive.ics.uci.edu/dataset/611/hierarchical+sales+data}
}
\keyword{datasets}
