% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_set_tidy}
\alias{data_set_tidy}
\title{Data set tidy}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 212164 rows and 5 columns.
}
\usage{
data_set_tidy
}
\description{
Data reshaped to be more "tidy"
}
\keyword{datasets}
