% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_all_items.R
\name{fit_all_items}
\alias{fit_all_items}
\title{Fit PAR Model to All Items in a Dataset}
\usage{
fit_all_items(item_list, q = 1)
}
\arguments{
\item{item_list}{A list of data objects (each with $y and $x)}

\item{q}{Number of lags to use in the model}
}
\value{
A list of model fits (each a list of $beta, $gamma, $loglik, etc.)
}
\description{
Loops through a list of item-specific data and fits the Poisson autoregressive model.
}
