% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_all_items.R
\name{fit_par_bfgs}
\alias{fit_par_bfgs}
\title{Fit PAR Model Using BFGS Optimization}
\usage{
fit_par_bfgs(spec, global_tol = 0.1, verbose = FALSE, maxIter = 1000)
}
\arguments{
\item{spec}{Model specification (output from \code{make_par_model_spec()})}

\item{global_tol}{Tolerance for convergence of the global optimization}

\item{verbose}{Verbosity level (0 = silent, 1 = progress, 2 = detailed)}

\item{maxIter}{Maximum number of iterations for the optimization}
}
\value{
A list containing the fitted model parameters
}
\description{
Fits a PAR model using the BFGS optimization algorithm.
}
