% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_model.R
\name{fit_par_grad_descent}
\alias{fit_par_grad_descent}
\title{Fit PAR Model Using Gradient Descent}
\usage{
fit_par_grad_descent(
  y,
  x,
  q = 1,
  initial_vals = NULL,
  lr = 1e-04,
  maxIter = 5000,
  tol = 1e-08,
  return_allIters = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{y}{Vector of counts}

\item{x}{Matrix of covariates (T × p). Include column of ones for intercept}

\item{q}{Number of lags}

\item{initial_vals}{Default NULL. Can be specified as a named list of parameter starting values, e.g. list(beta = rep(0,q), gamma = rep(0,p))}

\item{lr}{Learning rate for gradient descent (default = 1e-4)}

\item{maxIter}{Maximum iterations for gradient descent (default 5000)}

\item{tol}{Tolerance for gradient descent (default 1e-8)}

\item{return_allIters}{Default FALSE. Set TRUE to return all iterations from gradient descent}

\item{verbose}{Default FALSE. Set TRUE to print iterations while running}
}
\value{
A list with estimated parameters, log-likelihood, and convergence measure (epsilon)
}
\description{
Fit a non-hierarchical PAR model using maximum likelihood via projected gradient descent algorithm
}
