% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_all_items.R
\name{fit_par_item}
\alias{fit_par_item}
\title{Prepare and Fit PAR Model for a Single Item using Stan}
\usage{
fit_par_item(
  df,
  q = 1,
  stan_file = "inst/stan/par_item.stan",
  iter = 1000,
  chains = 4,
  seed = 123
)
}
\arguments{
\item{df}{A data.frame with columns \code{y}, \code{time}, and covariates (e.g., \code{promo})}

\item{q}{Number of AR lags}

\item{stan_file}{Path to Stan file (default = "inst/stan/par_item.stan")}

\item{iter}{Number of iterations (default = 1000)}

\item{chains}{Number of MCMC chains (default = 4)}

\item{seed}{Random seed}
}
\value{
A cmdstanr fit object
}
\description{
Prepare and Fit PAR Model for a Single Item using Stan
}
