% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_model.R
\name{fit_par_mcmc}
\alias{fit_par_mcmc}
\title{Fit PAR Model Using MCMC}
\usage{
fit_par_mcmc(
  y,
  x,
  q = 1,
  mcmc_iter = 5000,
  burn_in = 0.5,
  hyperparams = NULL,
  proposal_sd = 0.05,
  return_mcmc = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{y}{Vector of counts}

\item{x}{Matrix of covariates (T × p). Include column of ones for intercept}

\item{q}{Number of lags}

\item{mcmc_iter}{Number of iterations for MCMC. Default 5000}

\item{burn_in}{Proportion of burn-in samples for computing parameter estimates (default 0.5)}

\item{hyperparams}{Default NULL. Can be specified as named list of hyperparameters}

\item{proposal_sd}{Default 0.05. Standard deviation for Metropolis-Hasting proposal density}

\item{return_mcmc}{Default TRUE. Set FALSE to only return point estimates without MCMC samples}

\item{verbose}{Default FALSE. Set TRUE to print iterations while running}
}
\value{
A list with estimated parameters, log-likelihood, and (optionally) MCMC samples
}
\description{
Fit a non-hierarchical PAR model using MCMC (Metropolis-within-Gibbs sampler)
}
