% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar_model.R
\name{fit_par_mcmc_allitems}
\alias{fit_par_mcmc_allitems}
\title{Fit PAR Model for all items Using MCMC}
\usage{
fit_par_mcmc_allitems(
  Y,
  X,
  q = 1,
  mcmc_iter = 2000,
  burn_in = 0.5,
  hyperparams = NULL,
  proposal_sd = 0.05,
  return_mcmc = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{Y}{Matrix of counts (T x N)}

\item{X}{Matrix of covariates (T × N)}

\item{q}{Number of lags}

\item{mcmc_iter}{Number of iterations for MCMC. Default 2000}

\item{burn_in}{Proportion of burn-in samples for computing parameter estimates (default 0.5)}

\item{hyperparams}{Default NULL. Can be specified as named list of hyperparameters}

\item{proposal_sd}{Default 0.05. Standard deviation for Metropolis-Hasting proposal density}

\item{return_mcmc}{Default TRUE. Set FALSE to only return point estimates without MCMC samples}

\item{verbose}{Default FALSE. Set TRUE to print iterations while running}
}
\value{
A list with estimated parameters, log-likelihood, and (optionally) list of MCMC samples for each item
}
\description{
Fit a hierarchical multi-item Poisson Vector AR model using MCMC (Metropolis-within-Gibbs sampler)
}
