% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_model.R
\name{fit_par_optim}
\alias{fit_par_optim}
\title{Fit PAR Model Using Optim}
\usage{
fit_par_optim(y, x, q = 1)
}
\arguments{
\item{y}{Vector of counts}

\item{x}{Matrix of covariates (T × (q + p))}

\item{q}{Number of lags}
}
\value{
A list with estimated parameters and log-likelihood
}
\description{
Fit a non-hierarchical PAR model using maximum likelihood and optim().
}
\seealso{
\code{fit_par_bfgs()}
}
