% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_all_items.R
\name{fit_vector_par_stan}
\alias{fit_vector_par_stan}
\title{Fit the Vector PAR model in Stan}
\usage{
fit_vector_par_stan(
  stan_data,
  stan_file = "inst/stan/vector_par.stan",
  iter = 1000,
  chains = 4,
  seed = 123
)
}
\arguments{
\item{stan_data}{A list from prepare_data_stan_vectorpar()}

\item{stan_file}{Path to the Stan file}

\item{iter}{Number of iterations}

\item{chains}{Number of chains}

\item{seed}{Random seed}
}
\value{
A cmdstanr fit object
}
\description{
Fit the Vector PAR model in Stan
}
