% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_poisson_model.r
\name{general_bfgs}
\alias{general_bfgs}
\title{perform bfgs similar to optim}
\usage{
general_bfgs(
  params,
  f,
  g,
  ...,
  H = diag(length(params)),
  max_iter = 1000,
  trace_mod = 1
)
}
\arguments{
\item{params}{initial parameters}

\item{f}{the log likelihood function to maximize}

\item{g}{the gradient function}

\item{...}{additional arguments to pass to f and g}

\item{H}{initial inv-Hessian matrix. Default is the
identity, but convergence is more performant with the
true inv-Hessian matrix.}

\item{max_iter}{maximum number of iterations}

\item{trace_mod}{trace modulus. 0 is no tracing, 1 is
tracing each step, 2 is tracing every 2 steps, etc.}

\item{step_size}{step size in gradient descent}
}
\value{
optimized parameters
}
\description{
perform bfgs similar to optim
}
