% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_helpers.R
\name{is_par_model_spec}
\alias{is_par_model_spec}
\title{check if an object is a par_model_spec object}
\usage{
is_par_model_spec(x)
}
\arguments{
\item{x}{an object to test}
}
\description{
check if an object is a par_model_spec object
}
