% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nar_model_likelihood.R
\name{ll}
\alias{ll}
\title{Log-Likelihood of y and x given parameters}
\usage{
ll(y, x, alpha, betas, gamma, sigma2)
}
\arguments{
\item{y}{response vector,}

\item{x}{indicator vector}

\item{alpha}{scalar value for intercept effect}

\item{betas}{beta vector for each lag time point}

\item{gamma}{scalar value for effect of \code{x}}

\item{sigma2}{sigma squared}
}
\value{
Log-likelihood value.
}
\description{
Log-Likelihood of y and x given parameters
}
\examples{
betas <- c(.1, .2, .3)
gamma <-  10
alpha <- 3
t <- 100
x <- rep(c(0, 0, 0, 0, 1, 1, 1, 1, 0, 0), 10)
A <- diag(t) - beta_mat(betas, t)
A_inv <- solve(A)
sigma2 <- 1
V <- sigma2 * (A_inv \%*\% t(A_inv))
mu <- A_inv \%*\% (alpha + (gamma * x))

y <- mvtnorm::rmvnorm(n = 1, mean = mu, sigma = V)

## log-likelihood on values sampled from our given parameters
ll(t(y), matrix(x, ncol = 1), alpha, betas, gamma, sigma2)
}
