% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nar_model_mle.r
\name{mle_nar_model}
\alias{mle_nar_model}
\alias{mle_alpha}
\alias{mle_gamma}
\alias{mle_sigma2}
\alias{mle_beta}
\alias{optim_mle_params}
\title{Maximum Likelihood Estimation for NAR Model Parameters}
\usage{
mle_alpha(y, x, betas, gamma)

mle_gamma(y, x, betas, alpha)

mle_sigma2(y, x, betas, alpha, gamma)

mle_beta(y, x, alpha, betas, gamma, index)

optim_mle_params(y, x, alpha, betas, gamma, sigma2, times = 100, trace_mod = 1)
}
\arguments{
\item{y}{a matrix of dimension t by n}

\item{x}{a matrix of dimension t by n}

\item{betas}{a vector of coefficients for the autoregressive terms}

\item{gamma}{a scalar coefficient for the promotion}

\item{alpha}{a scalar coefficient for the intercept term}

\item{index}{the index of the betas to maximize}

\item{sigma2}{a scalar coefficient for the variance}

\item{times}{The number of times to run the optimization algorithm}

\item{trace_mod}{A number indicating the modulus to report the progress of the optimization algorithm}
}
\value{
an estimate for the respective Parameter
}
\description{
This function estimates the parameters of a Normal Autoregressive (NAR) model using Maximum Likelihood Estimation (MLE).
}
