% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mod_promo_brand_item}
\alias{mod_promo_brand_item}
\title{heirarchical Model Promotion brand-item}
\format{
An object of class \code{list} of length 2.
}
\usage{
mod_promo_brand_item
}
\value{
A par_model_spec
}
\description{
results of a long running Model for hierarchical
sales data with promotion, brand, and item effects
}
\keyword{datasets}
