% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{model_em_all}
\alias{model_em_all}
\title{EM silencing of PROMO variable}
\format{
An object of class \code{rowwise_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 118 rows and 5 columns.
}
\usage{
model_em_all
}
\value{
A tibble object
}
\description{
results of estimating Promotion effectiveness
latent variable per item brand combination
}
\keyword{datasets}
