% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_model.R
\name{par_deviance}
\alias{par_deviance}
\title{PAR model deviation
This is the deviance function used for poisson models}
\usage{
par_deviance(Y, X, beta, gamma)
}
\arguments{
\item{Y}{Vector of counts (length T)}

\item{X}{Matrix of covariates of dimension T × p}

\item{beta}{Vector of length q (autoregressive weights)}

\item{gamma}{Vector of length p (covariate effects)}
}
\description{
PAR model deviation
This is the deviance function used for poisson models
}
