% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promo_em.r
\name{par_em_effective}
\alias{par_em_effective}
\title{Perform EM - Silence a covariate}
\usage{
par_em_effective(
  spec,
  col,
  subset = NULL,
  tol = 0.001,
  max_iter = 100,
  show_plots = FALSE
)
}
\arguments{
\item{spec}{A model specification object.}

\item{col}{The column name of the covariate to silence.}

\item{subset}{A logical expression to subset the data.}

\item{tol}{The tolerance for convergence.}

\item{max_iter}{The maximum number of iterations.}

\item{show_plots}{Whether to print diagnostic information.}
}
\value{
a list of em model (spec), orignal model, z embeddings, and loglikelihood
}
\description{
Perform EM - Silence a covariate
}
\examples{
model_spec <- dplyr::filter(
     data_set_tidy,
     brand \%in\% c("B1", "B3"),
     item \%in\% c("1", "3")) |>
par_model_mat(
 formula = QTY ~ PROMO*brand*item,
 time = DATE,
 nlag = 4,
 groups = c(brand, item)
)
model_spec
em_out <- par_em_effective(model_spec,
   col = PROMO,
   subset = PROMO == 1,
   tol = 1e-6,
   show_plots = TRUE)
if (require("ggplot2")) {
  plot_data <- dplyr::ungroup(em_out$spec$.data) |>
     dplyr::mutate(em_mt = get_par_mt(em_out$spec),
            mt_original = get_par_mt(em_out$original_spec))
  ggplot(plot_data, aes(DATE)) +
    geom_line(aes(y = QTY, color = factor("Original Data"))) +
    geom_line(aes(y = mt_original, color = factor("Original Model")),
              linetype = "dotted", alpha = 0.8) +
    geom_rug(data = ~subset(.x, PROMO == 1)) +
    geom_line(aes(y = em_mt, color = factor("EM Model")),
              linetype = "dotted", alpha = 0.5) +
    facet_grid(cols = vars(brand), rows = vars(item)) +
    scale_color_manual(values = c("Original Data" = "grey",
                                   "Original Model" = "steelblue",
                                   "EM Model" = "red")) +
    theme_classic() +
    labs(color = "Model") +
    theme(legend.position = c(.8, .8))
}
}
