% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_model.R
\name{par_model_mat}
\alias{par_model_mat}
\title{Design matrix for analysis}
\usage{
par_model_mat(data, formula, time = NULL, nlag = 1, groups = NULL)
}
\arguments{
\item{data}{Data frame containing the data.}

\item{formula}{Formula specifying the model.}

\item{time}{Time variable.}

\item{nlag}{Number of lags.}

\item{groups}{Grouping variable.}
}
\description{
Design matrix for analysis
}
