% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data_allitems}
\alias{prepare_data_allitems}
\title{Prepare Raw Dataset for Time Series Modeling}
\usage{
prepare_data_allitems(data)
}
\arguments{
\item{data}{A raw dataframe with columns DATE and multiple columns starting with QTY and PROMO.}
}
\value{
A list containing time series across all items for sales outcome (Y) and promotions covariate (X), group/brand indices (G), items, and number of items and groups
}
\description{
Adds item/brand indices, arranges by time, and computes lags needed for PAR model.
}
