% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_helpers.R
\name{prepare_data_stan_vectorpar}
\alias{prepare_data_stan_vectorpar}
\title{Prepare Stan Data for Vector PAR Model}
\usage{
prepare_data_stan_vectorpar(data, q = 1)
}
\arguments{
\item{data}{Long-format tibble from simulate_vector_par()}

\item{q}{Number of lags}
}
\value{
A list ready to pass to Stan
}
\description{
Prepare Stan Data for Vector PAR Model
}
