% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_par.R
\name{simulate_vector_par}
\alias{simulate_vector_par}
\title{Simulate Hierarchical Vector Poisson Autoregressive (PAR) Time Series}
\usage{
simulate_vector_par(
  n_items = 30,
  n_brands = 3,
  T = 100,
  q = 1,
  p = 1,
  promo_prob = 0.2,
  sigma_mu = 1,
  rho_g = 0.5,
  sigma_eta = 0.5,
  seed = NULL
)
}
\arguments{
\item{n_items}{Number of items}

\item{n_brands}{Number of brands}

\item{T}{Number of timepoints}

\item{q}{Number of lags}

\item{p}{Number of covariates (default = 1)}

\item{promo_prob}{Probability of promotion being on}

\item{sigma_mu}{SD for brand-level mean of gamma}

\item{rho_g}{SD for item-level gamma noise}

\item{sigma_eta}{SD for item-level intercepts}

\item{seed}{Random seed}
}
\value{
A tibble with simulated long-format data
}
\description{
Simulate Hierarchical Vector Poisson Autoregressive (PAR) Time Series
}
