% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{split_by_item}
\alias{split_by_item}
\title{Split Preprocessed Data into Per-Item Lists}
\usage{
split_by_item(data, covariates = "PROMO")
}
\arguments{
\item{data}{A preprocessed dataframe (output of prepare_item_data)}

\item{covariates}{Character vector of covariate column names}
}
\value{
A list of item-specific lists each containing y and x
}
\description{
After using \code{prepare_item_data()}, split into list of per-item data with y and X matrix.
}
