% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{train_bayes_pvar}
\alias{train_bayes_pvar}
\title{Bayesian PAR/PVAR model training}
\usage{
train_bayes_pvar(
  data,
  ptrain = 0.8,
  model_type = "par",
  q = 1,
  mcmc_iter = 2000,
  verbose = TRUE
)
}
\arguments{
\item{data}{Preprocessed list of multi-item time series matrices, obtained from 'prepare_data_allitems' function on raw dataset}

\item{ptrain}{Proportion between 0-1 specifying the \% of timepoints to be used for model training. Default of 0.8}

\item{model_type}{String: Specify 'par' to fit PAR model, 'pvar' to fit PVAR model, or 'both' to fit both. If else, fits PAR model for each item. Default 'par'}

\item{q}{Integer lag for autoregressive terms. Default of 1}

\item{mcmc_iter}{Integer number of MCMC iterations for PAR/PVAR model fit. Default of 2000}

\item{verbose}{Set TRUE to return progress of model fitting/MCMC. Default TRUE}
}
\value{
List containing model fits and MCMC samples for specified model(s)
}
\description{
Fits the Bayesian PAR (single item at a time) and/or PVAR (multi-item/hierarchical) model(s) on training set, which is defined as first timepoints up to a specified cutoff
}
