% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{walk_promo_brand_item}
\alias{walk_promo_brand_item}
\title{bfgs walk heirarchical Model Promotion brand-item}
\format{
An object of class \code{character} of length 347.
}
\usage{
walk_promo_brand_item
}
\value{
A par_model_spec
}
\description{
verbose output of results of a long running Model for hierarchical
sales data with promotion, brand, and item effects
}
\keyword{datasets}
