% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastasales-package.R
\docType{package}
\name{pastasales-package}
\alias{pastasales}
\alias{pastasales-package}
\title{BIOS735 Group Project (group 2)}
\description{
This package contains functions for analyzing pasta sales data.
}
\author{
\strong{Maintainer}: Justin Landis \email{jtlandis314@gmail.com}

Authors:
\itemize{
  \item Alec Reinhardt \email{alec.e.reinhardt@gmail.com}
  \item Abby Foes \email{agfoes@gmail.com}
  \item Zheng Lan \email{zhenglan0822@gmail.com}
  \item Yu Liu \email{haibara.yl02@gmail.com}
}

}
\keyword{internal}
